+++ src/helpers/rightsReserved.ts
